+++ test/integration/012_deprecation_tests/test_deprecations.py
@@ -104,3 +104,77 @@ class TestDbtProjectYamlV1Deprecation(BaseTestDeprecations):
         self.run_dbt(strict=False)
         expected = {'dbt-project-yaml-v1'}
         self.assertEqual(expected, deprecations.active_deprecations)
+
+
+class TestAdapterMacroDeprecation(BaseTestDeprecations):
+    @property
+    def models(self):
+        return self.dir('adapter-macro-models')
+
+    @property
+    def project_config(self):
+        return {
+            'config-version': 2,
+            'macro-paths': [self.dir('adapter-macro-macros')]
+        }
+
+    @use_profile('postgres')
+    def test_postgres_adapter_macro(self):
+        self.assertEqual(deprecations.active_deprecations, set())
+        self.run_dbt(strict=False)
+        expected = {'adapter-macro'}
+        self.assertEqual(expected, deprecations.active_deprecations)
+
+    @use_profile('postgres')
+    def test_postgres_adapter_macro_fail(self):
+        self.assertEqual(deprecations.active_deprecations, set())
+        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
+            self.run_dbt(strict=True)
+        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
+        assert 'The "adapter_macro" macro has been deprecated' in exc_str
+
+    @use_profile('redshift')
+    def test_redshift_adapter_macro(self):
+        self.assertEqual(deprecations.active_deprecations, set())
+        # picked up the default -> error
+        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
+            self.run_dbt(strict=False, expect_pass=False)
+        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
+        assert 'not allowed' in exc_str  # we saw the default macro
+
+
+class TestAdapterMacroDeprecationPackages(BaseTestDeprecations):
+    @property
+    def models(self):
+        return self.dir('adapter-macro-models-package')
+
+    @property
+    def project_config(self):
+        return {
+            'config-version': 2,
+            'macro-paths': [self.dir('adapter-macro-macros')]
+        }
+
+    @use_profile('postgres')
+    def test_postgres_adapter_macro_pkg(self):
+        self.assertEqual(deprecations.active_deprecations, set())
+        self.run_dbt(strict=False)
+        expected = {'adapter-macro'}
+        self.assertEqual(expected, deprecations.active_deprecations)
+
+    @use_profile('postgres')
+    def test_postgres_adapter_macro_pkg_fail(self):
+        self.assertEqual(deprecations.active_deprecations, set())
+        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
+            self.run_dbt(strict=True)
+        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
+        assert 'The "adapter_macro" macro has been deprecated' in exc_str
+
+    @use_profile('redshift')
+    def test_redshift_adapter_macro_pkg(self):
+        self.assertEqual(deprecations.active_deprecations, set())
+        # picked up the default -> error
+        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
+            self.run_dbt(strict=False, expect_pass=False)
+        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
+        assert 'not allowed' in exc_str  # we saw the default macro