+++ test/integration/026_aliases_test/macros/cast.sql
@@ -1,7 +1,7 @@
 
 
 {% macro string_literal(s) -%}
-  {{ adapter_macro('test.string_literal', s) }}
+  {{ adapter.dispatch('string_literal', packages=['test'])(s) }}
 {%- endmacro %}
 
 {% macro default__string_literal(s) %}